+++ include/Queue.h
@@ -5,13 +5,13 @@
 
 typedef struct { // Circular Queue
     genVec* arr;
-    size_t head;     // dequeue (head + 1) % capacity
-    size_t tail;     // enqueue   (head + size) % capacity
-    size_t size;
+    u32 head;     // dequeue (head + 1) % capacity
+    u32 tail;     // enqueue   (head + size) % capacity
+    u32 size;
 } __attribute__((aligned(32))) Queue;
 
 
-Queue* queue_create(size_t n, size_t data_size, genVec_delete_fn del_fn);
+Queue* queue_create(u32 n, u32 data_size, genVec_delete_fn del_fn);
 void queue_destroy(Queue* q);
 void queue_clear(Queue* q);
 
@@ -19,7 +19,7 @@ void enqueue(Queue* q, const u8* x);
 void dequeue(Queue* q, u8* out);
 void queue_peek(Queue* q, u8* peek);
 
-static inline size_t queue_size(Queue* q) {
+static inline u32 queue_size(Queue* q) {
     return q->size;
 }
