+++ include/default_functions.h
@@ -16,10 +16,10 @@ typedef uint8_t u8;
 ====================DEFAULT FUNCTIONS====================
 */
 // 32-bit FNV-1a (default hash)
-static size_t fnv1a_hash(const u8* bytes, size_t size) {
+static u32 fnv1a_hash(const u8* bytes, u32 size) {
     uint32_t hash = 2166136261U;  // FNV offset basis
 
-    for (size_t i = 0; i < size; i++) {
+    for (u32 i = 0; i < size; i++) {
         hash ^= bytes[i];   // XOR with current byte
         hash *= 16777619U;  // Multiply by FNV prime
     }
@@ -29,22 +29,22 @@ static size_t fnv1a_hash(const u8* bytes, size_t size) {
 
 
 // Default compare function
-static int default_compare(const u8* a, const u8* b, size_t size) 
+static int default_compare(const u8* a, const u8* b, u32 size) 
 {
     return memcmp(a, b, size);
 }
 
-static const size_t PRIMES[] = {
+static const u32 PRIMES[] = {
     17, 37, 79, 163, 331, 673, 1361, 2729, 5471, 10949, 
     21911, 43853, 87719, 175447, 350899, 701819, 1403641, 
     2807303, 5614657, 11229331, 22458671, 44917381, 89834777
 };
 
-static const size_t PRIMES_COUNT = sizeof(PRIMES) / sizeof(PRIMES[0]);
+static const u32 PRIMES_COUNT = sizeof(PRIMES) / sizeof(PRIMES[0]);
 
 // Find the next prime number larger than current
-static size_t next_prime(size_t current) {
-    for (size_t i = 0; i < PRIMES_COUNT; i++) {
+static u32 next_prime(u32 current) {
+    for (u32 i = 0; i < PRIMES_COUNT; i++) {
         if (PRIMES[i] > current) {
             return PRIMES[i];
         }
@@ -57,9 +57,9 @@ static size_t next_prime(size_t current) {
 }
 
 // Find the previous prime number smaller than current
-static size_t prev_prime(size_t current) {
+static u32 prev_prime(u32 current) {
     // Search backwards through prime table
-    for (size_t i = PRIMES_COUNT - 1; i >= 0; i--) {
+    for (u32 i = PRIMES_COUNT - 1; i >= 0; i--) {
         if (PRIMES[i] < current) {
             return PRIMES[i];
         }
