+++ include/map_setup.h
@@ -14,5 +14,5 @@ typedef enum {
 
 
 typedef void (*delete_fn)(u8* key); //optional (for string*, arrays, structs as keys, vals)
-typedef size_t (*custom_hash_fn)(const u8* key, size_t size);     // optional (for a specific usecase)
-typedef int (*compare_fn)(const u8* a, const u8* b, size_t size);   // optional: for compare operations
+typedef u32 (*custom_hash_fn)(const u8* key, u32 size);     // optional (for a specific usecase)
+typedef int (*compare_fn)(const u8* a, const u8* b, u32 size);   // optional: for compare operations
